gossmods_from_listpeerchannels: use correct type for cltv_delta.
Doesn't matter now, but will with the next change where we want to
pass a pointer.

+++ common/gossmods_listpeerchannels.h
@@ -35,7 +35,7 @@ struct gossmap_localmods *gossmods_from_listpeerchannels_(const tal_t *ctx,
 								     struct amount_msat spendable,
 								     struct amount_msat fee_base,
 								     u32 fee_proportional,
-								     u32 cltv_delta,
+								     u16 cltv_delta,
 								     bool enabled,
 								     const char *buf_,
 								     const jsmntok_t *chantok,
@@ -54,7 +54,7 @@ struct gossmap_localmods *gossmods_from_listpeerchannels_(const tal_t *ctx,
 							    struct amount_msat,	\
 							    struct amount_msat,	\
 							    u32,	\
-							    u32,	\
+							    u16,	\
 							    bool,	\
 							    const char *, \
 							    const jsmntok_t *), \
@@ -70,7 +70,7 @@ void gossmod_add_localchan(struct gossmap_localmods *mods,
 			   struct amount_msat spendable,
 			   struct amount_msat fee_base,
 			   u32 fee_proportional,
-			   u32 cltv_delta,
+			   u16 cltv_delta,
 			   bool enabled,
 			   const char *buf UNUSED,
 			   const jsmntok_t *chantok UNUSED,
